Split benchmarks into separate files
