--- crates/ecs_bench_suite/benches/benchmarks.rs
@@ -1,243 +1,0 @@
-use criterion::*;
-use ecs_bench_suite::*;
-
-fn bench_simple_insert(c: &mut Criterion) {
-    let mut group = c.benchmark_group("simple_insert");
-    #[cfg(feature = "bench-all-engines")]
-    {
-        group.bench_function("legion", |b| {
-            let mut bench = legion::simple_insert::Benchmark::new();
-            b.iter(move || bench.run());
-        });
-        group.bench_function("bevy", |b| {
-            let mut bench = bevy::simple_insert::Benchmark::new();
-            b.iter(move || bench.run());
-        });
-        group.bench_function("hecs", |b| {
-            let mut bench = hecs::simple_insert::Benchmark::new();
-            b.iter(move || bench.run());
-        });
-        group.bench_function("planck_ecs", |b| {
-            let mut bench = planck_ecs::simple_insert::Benchmark::new();
-            b.iter(move || bench.run());
-        });
-        group.bench_function("shipyard", |b| {
-            let mut bench = shipyard::simple_insert::Benchmark::new();
-            b.iter(move || bench.run());
-        });
-        group.bench_function("specs", |b| {
-            let mut bench = specs::simple_insert::Benchmark::new();
-            b.iter(move || bench.run());
-        });
-    }
-    group.bench_function("recs", |b| {
-        let mut bench = recs::simple_insert::Benchmark::new();
-        b.iter(move || bench.run());
-    });
-    group.bench_function("naive_vec", |b| {
-        let mut bench = naive::simple_insert::Benchmark::new();
-        b.iter(move || bench.run());
-    });
-}
-
-fn bench_simple_iter(c: &mut Criterion) {
-    let mut group = c.benchmark_group("simple_iter");
-    #[cfg(feature = "bench-all-engines")]
-    {
-        group.bench_function("legion", |b| {
-            let mut bench = legion::simple_iter::Benchmark::new();
-            b.iter(move || bench.run());
-        });
-        group.bench_function("legion (packed)", |b| {
-            let mut bench = legion_packed::simple_iter::Benchmark::new();
-            b.iter(move || bench.run());
-        });
-        group.bench_function("bevy", |b| {
-            let mut bench = bevy::simple_iter::Benchmark::new();
-            b.iter(move || bench.run());
-        });
-        group.bench_function("hecs", |b| {
-            let mut bench = hecs::simple_iter::Benchmark::new();
-            b.iter(move || bench.run());
-        });
-        group.bench_function("planck_ecs", |b| {
-            let mut bench = planck_ecs::simple_iter::Benchmark::new();
-            b.iter(move || bench.run());
-        });
-        group.bench_function("shipyard", |b| {
-            let mut bench = shipyard::simple_iter::Benchmark::new();
-            b.iter(move || bench.run());
-        });
-        group.bench_function("specs", |b| {
-            let mut bench = specs::simple_iter::Benchmark::new();
-            b.iter(move || bench.run());
-        });
-    }
-    group.bench_function("recs", |b| {
-        let mut bench = recs::simple_iter::Benchmark::new();
-        b.iter(move || bench.run());
-    });
-    group.bench_function("naive_vec", |b| {
-        let mut bench = naive::simple_iter::Benchmark::new();
-        b.iter(move || bench.run());
-    });
-}
-
-fn bench_frag_iter_bc(c: &mut Criterion) {
-    let mut group = c.benchmark_group("fragmented_iter");
-    #[cfg(feature = "bench-all-engines")]
-    {
-        group.bench_function("legion", |b| {
-            let mut bench = legion::frag_iter::Benchmark::new();
-            b.iter(move || bench.run());
-        });
-        group.bench_function("bevy", |b| {
-            let mut bench = bevy::frag_iter::Benchmark::new();
-            b.iter(move || bench.run());
-        });
-        group.bench_function("hecs", |b| {
-            let mut bench = hecs::frag_iter::Benchmark::new();
-            b.iter(move || bench.run());
-        });
-        group.bench_function("planck_ecs", |b| {
-            let mut bench = planck_ecs::frag_iter::Benchmark::new();
-            b.iter(move || bench.run());
-        });
-        group.bench_function("shipyard", |b| {
-            let mut bench = shipyard::frag_iter::Benchmark::new();
-            b.iter(move || bench.run());
-        });
-        group.bench_function("specs", |b| {
-            let mut bench = specs::frag_iter::Benchmark::new();
-            b.iter(move || bench.run());
-        });
-    }
-    group.bench_function("recs", |b| {
-        let mut bench = recs::frag_iter::Benchmark::new();
-        b.iter(move || bench.run());
-    });
-}
-
-fn bench_schedule(c: &mut Criterion) {
-    let mut group = c.benchmark_group("schedule");
-    #[cfg(feature = "bench-all-engines")]
-    {
-        group.bench_function("legion", |b| {
-            let mut bench = legion::schedule::Benchmark::new();
-            b.iter(move || bench.run());
-        });
-        group.bench_function("legion (packed)", |b| {
-            let mut bench = legion_packed::schedule::Benchmark::new();
-            b.iter(move || bench.run());
-        });
-        group.bench_function("bevy", |b| {
-            let mut bench = bevy::schedule::Benchmark::new();
-            b.iter(move || bench.run());
-        });
-        group.bench_function("planck_ecs", |b| {
-            let mut bench = planck_ecs::schedule::Benchmark::new();
-            b.iter(move || bench.run());
-        });
-        group.bench_function("shipyard", |b| {
-            let mut bench = shipyard::schedule::Benchmark::new();
-            b.iter(move || bench.run());
-        });
-        group.bench_function("specs", |b| {
-            let mut bench = specs::schedule::Benchmark::new();
-            b.iter(move || bench.run());
-        });
-    }
-    group.bench_function("recs", |b| {
-        let mut bench = recs::schedule::Benchmark::new();
-        b.iter(move || bench.run());
-    });
-}
-
-fn bench_heavy_compute(c: &mut Criterion) {
-    let mut group = c.benchmark_group("heavy_compute");
-    #[cfg(feature = "bench-all-engines")]
-    {
-        group.bench_function("legion", |b| {
-            let mut bench = legion::heavy_compute::Benchmark::new();
-            b.iter(move || bench.run());
-        });
-        group.bench_function("legion (packed)", |b| {
-            let mut bench = legion_packed::heavy_compute::Benchmark::new();
-            b.iter(move || bench.run());
-        });
-        group.bench_function("bevy", |b| {
-            let mut bench = bevy::heavy_compute::Benchmark::new();
-            b.iter(move || bench.run());
-        });
-        group.bench_function("hecs", |b| {
-            let mut bench = hecs::heavy_compute::Benchmark::new();
-            b.iter(move || bench.run());
-        });
-        group.bench_function("shipyard", |b| {
-            let mut bench = shipyard::heavy_compute::Benchmark::new();
-            b.iter(move || bench.run());
-        });
-        group.bench_function("specs", |b| {
-            let mut bench = specs::heavy_compute::Benchmark::new();
-            b.iter(move || bench.run());
-        });
-    }
-    group.bench_function("recs", |b| {
-        let mut bench = recs::heavy_compute::Benchmark::new();
-        b.iter(move || bench.run());
-    });
-    group.bench_function("naive_vec", |b| {
-        let mut bench = naive::heavy_compute::Benchmark::new();
-        b.iter(move || bench.run());
-    });
-}
-
-fn bench_add_remove(c: &mut Criterion) {
-    let mut group = c.benchmark_group("add_remove_component");
-    #[cfg(feature = "bench-all-engines")]
-    {
-        group.bench_function("legion", |b| {
-            let mut bench = legion::add_remove::Benchmark::new();
-            b.iter(move || bench.run());
-        });
-        group.bench_function("hecs", |b| {
-            let mut bench = hecs::add_remove::Benchmark::new();
-            b.iter(move || bench.run());
-        });
-        group.bench_function("planck_ecs", |b| {
-            let mut bench = planck_ecs::add_remove::Benchmark::new();
-            b.iter(move || bench.run());
-        });
-        group.bench_function("shipyard", |b| {
-            let mut bench = shipyard::add_remove::Benchmark::new();
-            b.iter(move || bench.run());
-        });
-        group.bench_function("specs", |b| {
-            let mut bench = specs::add_remove::Benchmark::new();
-            b.iter(move || bench.run());
-        });
-        group.bench_function("bevy", |b| {
-            let mut bench = bevy::add_remove::Benchmark::new();
-            b.iter(move || bench.run());
-        });
-    }
-    group.bench_function("recs", |b| {
-        let mut bench = recs::add_remove::Benchmark::new();
-        b.iter(move || bench.run());
-    });
-    group.bench_function("naive_vec", |b| {
-        let mut bench = naive::add_remove::Benchmark::new();
-        b.iter(move || bench.run());
-    });
-}
-
-criterion_group!(
-    all_benches,
-    bench_simple_insert,
-    bench_simple_iter,
-    bench_frag_iter_bc,
-    bench_schedule,
-    bench_heavy_compute,
-    bench_add_remove,
-);
-criterion_main!(all_benches,);
